refactor change editor button
