+++ redisinsight/ui/src/pages/browser/modules/key-details/components/change-editor-type-button/index.ts
@@ -1,0 +1,4 @@
+import ChangeEditorTypeButton from './change-editor-type-button'
+
+export { useChangeEditorType } from './useChangeEditorType'
+export default ChangeEditorTypeButton
